move out bitbucket into a separate module too
