+++ .changeset/silent-mugs-roll.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-catalog-backend-module-bitbucket': minor
+---
+
+Added package, moving out Bitbucket specific functionality from the catalog-backend
