+++ plugins/catalog-backend-module-bitbucket/.eslintrc.js
