+++ plugins/catalog-backend-module-bitbucket/README.md
@@ -1,0 +1,8 @@
+# Catalog Backend Module for Bitbucket
+
+This is an extension module to the plugin-catalog-backend plugin, providing extensions targeted at Bitbucket offerings.
+
+## Getting started
+
+See [Backstage documentation](https://backstage.io/docs/integrations/bitbucket/discovery) for details on how to install
+and configure the plugin.
