+++ plugins/catalog-backend-module-bitbucket/api-report.md
@@ -1,0 +1,49 @@
+## API Report File for "@backstage/plugin-catalog-backend-module-bitbucket"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { BitbucketIntegration } from '@backstage/integration';
+import { CatalogProcessor } from '@backstage/plugin-catalog-backend';
+import { CatalogProcessorEmit } from '@backstage/plugin-catalog-backend';
+import { CatalogProcessorResult } from '@backstage/plugin-catalog-backend';
+import { Config } from '@backstage/config';
+import { LocationSpec } from '@backstage/plugin-catalog-backend';
+import { Logger } from 'winston';
+import { ScmIntegrationRegistry } from '@backstage/integration';
+
+// @public (undocumented)
+export class BitbucketDiscoveryProcessor implements CatalogProcessor {
+  constructor(options: {
+    integrations: ScmIntegrationRegistry;
+    parser?: (options: {
+      integration: BitbucketIntegration;
+      target: string;
+      presence?: 'optional' | 'required';
+      logger: Logger;
+    }) => AsyncIterable<CatalogProcessorResult>;
+    logger: Logger;
+  });
+  // (undocumented)
+  static fromConfig(
+    config: Config,
+    options: {
+      parser?: (options: {
+        integration: BitbucketIntegration;
+        target: string;
+        presence?: 'optional' | 'required';
+        logger: Logger;
+      }) => AsyncIterable<CatalogProcessorResult>;
+      logger: Logger;
+    },
+  ): BitbucketDiscoveryProcessor;
+  // (undocumented)
+  getProcessorName(): string;
+  // (undocumented)
+  readLocation(
+    location: LocationSpec,
+    _optional: boolean,
+    emit: CatalogProcessorEmit,
+  ): Promise<boolean>;
+}
+```
