+++ plugins/catalog-backend-module-bitbucket/package.json
@@ -1,0 +1,56 @@
+{
+  "name": "@backstage/plugin-catalog-backend-module-bitbucket",
+  "description": "A Backstage catalog backend module that helps integrate towards Bitbucket",
+  "version": "0.0.0",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "license": "Apache-2.0",
+  "private": false,
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.cjs.js",
+    "types": "dist/index.d.ts"
+  },
+  "backstage": {
+    "role": "backend-plugin-module"
+  },
+  "homepage": "https://backstage.io",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/backstage/backstage",
+    "directory": "plugins/catalog-backend-module-bitbucket"
+  },
+  "keywords": [
+    "backstage"
+  ],
+  "scripts": {
+    "build": "backstage-cli package build",
+    "lint": "backstage-cli package lint",
+    "test": "backstage-cli package test",
+    "prepack": "backstage-cli package prepack",
+    "postpack": "backstage-cli package postpack",
+    "clean": "backstage-cli package clean",
+    "start": "backstage-cli package start"
+  },
+  "dependencies": {
+    "@backstage/backend-common": "^0.12.0",
+    "@backstage/catalog-model": "^0.12.0",
+    "@backstage/config": "^0.1.15",
+    "@backstage/errors": "^0.2.2",
+    "@backstage/integration": "^0.8.0",
+    "@backstage/plugin-catalog-backend": "^0.23.0",
+    "@backstage/types": "^0.1.3",
+    "lodash": "^4.17.21",
+    "msw": "^0.35.0",
+    "node-fetch": "^2.6.7",
+    "winston": "^3.2.1"
+  },
+  "devDependencies": {
+    "@backstage/backend-test-utils": "^0.1.20",
+    "@backstage/cli": "^0.15.0",
+    "@types/lodash": "^4.14.151"
+  },
+  "files": [
+    "dist"
+  ]
+}
