+++ plugins/catalog-backend-module-bitbucket/src/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Backstage Authors
+ * Copyright 2020 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,8 +14,10 @@
  * limitations under the License.
  */
 
-export { BitbucketClient, paginated, paginated20 } from './client';
-export { defaultRepositoryParser } from './BitbucketRepositoryParser';
-export type { PagedResponse, PagedResponse20 } from './client';
-export type { BitbucketRepository, BitbucketRepository20 } from './types';
-export type { BitbucketRepositoryParser } from './BitbucketRepositoryParser';
+/**
+ * A Backstage catalog backend module that helps integrate towards GitLab
+ *
+ * @packageDocumentation
+ */
+
+export { BitbucketDiscoveryProcessor } from './BitbucketDiscoveryProcessor';
