+++ plugins/catalog-backend-module-bitbucket/src/lib/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2021 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,5 +14,8 @@
  * limitations under the License.
  */
 
-export { BitbucketDiscoveryProcessor } from './BitbucketDiscoveryProcessor';
-export type { BitbucketRepositoryParser } from './lib';
+export { defaultRepositoryParser } from './BitbucketRepositoryParser';
+export type { BitbucketRepositoryParser } from './BitbucketRepositoryParser';
+export { BitbucketClient, paginated, paginated20 } from './client';
+export type { PagedResponse, PagedResponse20 } from './client';
+export type { BitbucketRepository, BitbucketRepository20 } from './types';
