+++ plugins/catalog-backend-module-bitbucket/src/lib/types.ts
