+++ plugins/catalog-backend-module-gitlab/.eslintrc.js
@@ -1,3 +1,1 @@
-module.exports = {
-  extends: [require.resolve('@backstage/cli/config/eslint.backend')],
-};
+module.exports = require('@backstage/cli/config/eslint-factory')(__dirname);
